+++ src/main/java/net/geforcemods/securitycraft/api/ICustomizable.java
@@ -24,9 +24,10 @@ public interface ICustomizable {
 	/**
 	 * Called whenever an {@link Option} in this TileEntity changes its value
 	 *
+	 * @param <T> The type the Option stores
 	 * @param option The changed Option
 	 */
-	public default void onOptionChanged(Option<?> option) {
+	public default <T> void onOptionChanged(Option<T> option) {
 		getTheBlockEntity().setChanged();
 	}