+++ src/main/java/net/geforcemods/securitycraft/blockentities/InventoryScannerBlockEntity.java
@@ -431,39 +431,41 @@ public class InventoryScannerBlockEntity extends DisguisableBlockEntity implemen
 	}
 
 	@Override
-	public void onOptionChanged(Option<?> option) {
-		if (option.getName().equals("horizontal")) {
-			BooleanOption bo = (BooleanOption) option;
+	public <T> void onOptionChanged(Option<T> option) {
+		if (option instanceof BooleanOption bo) {
+			if (option == horizontal) {
+				modifyFields((offsetPos, state) -> level.setBlockAndUpdate(offsetPos, state.setValue(InventoryScannerFieldBlock.HORIZONTAL, bo.get())), connectedScanner -> connectedScanner.setHorizontal(bo.get()));
+				level.setBlockAndUpdate(worldPosition, getBlockState().setValue(InventoryScannerBlock.HORIZONTAL, bo.get()));
+			}
+			else if (option == solidifyField) {
+				InventoryScannerBlockEntity connectedScanner = InventoryScannerBlock.getConnectedInventoryScanner(level, worldPosition);
 
-			modifyFields((offsetPos, state) -> level.setBlockAndUpdate(offsetPos, state.setValue(InventoryScannerFieldBlock.HORIZONTAL, bo.get())), connectedScanner -> connectedScanner.setHorizontal(bo.get()));
-			level.setBlockAndUpdate(worldPosition, getBlockState().setValue(InventoryScannerBlock.HORIZONTAL, bo.get()));
-		}
-		else if (option.getName().equals("solidifyField")) {
-			InventoryScannerBlockEntity connectedScanner = InventoryScannerBlock.getConnectedInventoryScanner(level, worldPosition);
+				if (connectedScanner != null)
+					connectedScanner.setSolidifyField(bo.get());
+			}
+			else if (option == disabled) {
+				if (!bo.get())
+					InventoryScannerBlock.checkAndPlaceAppropriately(level, worldPosition, true);
+				else
+					modifyFields((offsetPos, state) -> level.destroyBlock(offsetPos, false), connectedScanner -> connectedScanner.setDisabled(true));
+			}
+			else if (option == ignoreOwner) {
+				InventoryScannerBlockEntity connectedScanner = InventoryScannerBlock.getConnectedInventoryScanner(level, worldPosition);
 
-			if (connectedScanner != null)
-				connectedScanner.setSolidifyField(((BooleanOption) option).get());
-		}
-		else if (option.getName().equals("disabled")) {
-			BooleanOption bo = (BooleanOption) option;
-
-			if (!bo.get())
-				InventoryScannerBlock.checkAndPlaceAppropriately(level, worldPosition, true);
+				if (connectedScanner != null)
+					connectedScanner.setIgnoresOwner(bo.get());
+			}
 			else
-				modifyFields((offsetPos, state) -> level.destroyBlock(offsetPos, false), connectedScanner -> connectedScanner.setDisabled(true));
+				throw new UnsupportedOperationException("Unhandled option synchronization in inventory scanner! " + option.getName());
 		}
-		else if (option.getName().equals("ignoreOwner")) {
+		else if (option instanceof IntOption io && option == signalLength) {
 			InventoryScannerBlockEntity connectedScanner = InventoryScannerBlock.getConnectedInventoryScanner(level, worldPosition);
 
 			if (connectedScanner != null)
-				connectedScanner.setIgnoresOwner(((BooleanOption) option).get());
+				connectedScanner.setSignalLength(io.get());
 		}
-		else if (option.getName().equals("signalLength")) {
-			InventoryScannerBlockEntity connectedScanner = InventoryScannerBlock.getConnectedInventoryScanner(level, worldPosition);
-
-			if (connectedScanner != null)
-				connectedScanner.setSignalLength(((IntOption) option).get());
-		}
+		else
+			throw new UnsupportedOperationException("Unhandled option synchronization in inventory scanner! " + option.getName());
 
 		super.onOptionChanged(option);
 	}